pyproject.toml
[project]
name = "deployment-yamls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.1.0",
]
